# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Film": {
				"properties": {
					"film_id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"release_year": {
						"type": "number",
						"format": "double"
					},
					"language_id": {
						"type": "number",
						"format": "double"
					},
					"original_language_id": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"rental_duration": {
						"type": "number",
						"format": "double"
					},
					"rental_rate": {
						"type": "number",
						"format": "double"
					},
					"length": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"replacement_cost": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "string"
					},
					"special_features": {
						"type": "string",
						"nullable": true
					},
					"last_update": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"film_id",
					"title",
					"description",
					"release_year",
					"language_id",
					"original_language_id",
					"rental_duration",
					"rental_rate",
					"length",
					"replacement_cost",
					"rating",
					"special_features",
					"last_update"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedFilms": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Film"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"pageSize"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Actor": {
				"properties": {
					"actor_id": {
						"type": "number",
						"format": "double"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"last_update": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"actor_id",
					"first_name",
					"last_name",
					"last_update"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedActors": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/Actor"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"pageSize"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ActorCreationParams": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					}
				},
				"required": [
					"first_name",
					"last_name"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "api",
		"version": "1.0.0",
		"description": "joel",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "joel"
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"paths": {
...
		"/films/actor/{actorId}": {
			"get": {
				"operationId": "GetFilmsByActor",
				"responses": {
					"200": {
						"description": "Films par acteur récupérés avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedFilms"
								}
							}
						}
					}
				},
				"description": "Récupère les films par acteur",
				"tags": [
					"Films"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant de l'acteur",
						"in": "path",
						"name": "actorId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Numéro de la page",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/actors": {
			"get": {
				"operationId": "GetActors",
				"responses": {
					"200": {
						"description": "Liste des acteurs récupérée avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaginatedActors"
								}
							}
						}
					}
				},
				"description": "Récupère la liste de tous les acteurs avec pagination",
				"tags": [
					"Actors"
				],
				"security": [],
				"parameters": [
					{
						"description": "Numéro de la page (commence à 1)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateActor",
				"responses": {
					"201": {
						"description": "Acteur créé avec succès",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Actor"
								}
							}
						}
					}
				},
				"description": "Crée un nouvel acteur",
				"tags": [
					"Actors"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Données de l'acteur à créer",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ActorCreationParams",
								"description": "Données de l'acteur à créer"
							}
						}
					}
				}
			}
		},
		"/actors/{id}": {
			"get": {
				"operationId": "GetActorById",
				"responses": {
					"200": {
						"description": "Acteur récupéré avec succès",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Actor"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Récupère un acteur par son identifiant",
				"tags": [
					"Actors"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant unique de l'acteur",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateActor",
				"responses": {
					"200": {
						"description": "Acteur mis à jour avec succès",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Actor"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Met à jour un acteur existant",
				"tags": [
					"Actors"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Identifiant de l'acteur à mettre à jour",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Nouvelles données de l'acteur",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ActorCreationParams",
								"description": "Nouvelles données de l'acteur"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteActor",
				"responses": {
					"204": {
						"description": "Acteur supprimé avec succès",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Supprime un acteur",
				"tags": [
					"Actors"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Identifiant de l'acteur à supprimer",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/actors/{id}/films": {
			"get": {
				"operationId": "GetActorFilms",
				"responses": {
					"200": {
						"description": "Films de l'acteur récupérés avec succès",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"pageSize": {
											"type": "number",
											"format": "double"
										},
										"page": {
											"type": "number",
											"format": "double"
										},
										"total": {
											"type": "number",
											"format": "double"
										},
										"data": {
											"items": {},
											"type": "array"
										}
									},
									"required": [
										"pageSize",
										"page",
										"total",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Récupère les films d'un acteur spécifique",
				"tags": [
					"Actors"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identifiant de l'acteur",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Numéro de la page",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Nombre d'éléments par page",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	}
}